Fixed manual entry to manage page
Can no longer change the URL to get into the manage page. Manually doing so will create a null user ID. This is checked for and returns the user back to the sign in page. Other code such as Add Contact and Search Contact are being worked on, but not finished yet.